clear up language, style guide, new graph
Slight change to make it that the description is referencing to card deck clear.
More significant was numerous updates to try to bring to conformity with style guide, add all the references, and add one last Venn Diagram. Significantly also changes default from slider to typing in the values which I think works much better.